+++ config/max-lines-baseline.txt
@@ -500,7 +500,6 @@ src/agents/tool-loop-detection.test.ts
 src/agents/tool-search.test.ts
 src/agents/tools/computer-tool.ts
 src/agents/tools/cron-tool.test.ts
-src/agents/tools/cron-tool.ts
 src/agents/tools/image-generate-tool.test.ts
 src/agents/tools/image-generate-tool.ts
 src/agents/tools/image-tool.test.ts
@@ -687,14 +686,11 @@ src/config/zod-schema.core.ts
 src/context-engine/context-engine.test.ts
 src/context-engine/registry.ts
 src/cron/isolated-agent/delivery-dispatch.double-announce.test.ts
-src/cron/isolated-agent/delivery-dispatch.ts
 src/cron/isolated-agent/delivery-target.test.ts
 src/cron/isolated-agent/run-executor.ts
 src/cron/isolated-agent/run.message-tool-policy.test.ts
 src/cron/isolated-agent/run.test-harness.ts
-src/cron/isolated-agent/run.ts
 src/cron/service.jobs.test.ts
-src/cron/service/jobs.ts
 src/cron/service/ops.test.ts
 src/cron/service/timer.regression.test.ts
 src/cron/store.test.ts
