+++ src/agents/tools/cron-tool-context.ts
@@ -1,0 +1,92 @@
+/** Reminder-context projection for cron tool job creation. */
+import { getRuntimeConfig } from "../../config/config.js";
+import { extractTextFromChatContent } from "../../shared/chat-content.js";
+import { truncateUtf16Safe } from "../../utils.js";
+import { REMINDER_CONTEXT_MESSAGES_MAX } from "./cron-tool-schema.js";
+import type { ChatMessage, GatewayToolCaller } from "./cron-tool.types.js";
+import type { GatewayCallOptions } from "./gateway.js";
+import { resolveInternalSessionKey, resolveMainSessionAlias } from "./sessions-helpers.js";
+
+const REMINDER_CONTEXT_PER_MESSAGE_MAX = 220;
+const REMINDER_CONTEXT_TOTAL_MAX = 700;
+export const REMINDER_CONTEXT_MARKER = "\n\nRecent context:\n";
+
+export function stripExistingContext(text: string) {
+  const index = text.indexOf(REMINDER_CONTEXT_MARKER);
+  if (index === -1) {
+    return text;
+  }
+  return text.slice(0, index).trim();
+}
+
+function truncateText(input: string, maxLen: number) {
+  if (input.length <= maxLen) {
+    return input;
+  }
+  const truncated = truncateUtf16Safe(input, Math.max(0, maxLen - 3)).trimEnd();
+  return `${truncated}...`;
+}
+
+function extractMessageText(message: ChatMessage): { role: string; text: string } | null {
+  const role = typeof message.role === "string" ? message.role : "";
+  if (role !== "user" && role !== "assistant") {
+    return null;
+  }
+  const text = extractTextFromChatContent(message.content);
+  return text ? { role, text } : null;
+}
+
+export async function buildReminderContextLines(params: {
+  agentSessionKey?: string;
+  gatewayOpts: GatewayCallOptions;
+  contextMessages: number;
+  callGatewayTool: GatewayToolCaller;
+}) {
+  const maxMessages = Math.min(
+    REMINDER_CONTEXT_MESSAGES_MAX,
+    Math.max(0, Math.floor(params.contextMessages)),
+  );
+  if (maxMessages <= 0) {
+    return [];
+  }
+  const sessionKey = params.agentSessionKey?.trim();
+  if (!sessionKey) {
+    return [];
+  }
+  const cfg = getRuntimeConfig();
+  const { mainKey, alias } = resolveMainSessionAlias(cfg);
+  const resolvedKey = resolveInternalSessionKey({ key: sessionKey, alias, mainKey });
+  try {
+    const res = await params.callGatewayTool<{ messages: Array<unknown> }>(
+      "chat.history",
+      params.gatewayOpts,
+      {
+        sessionKey: resolvedKey,
+        limit: maxMessages,
+      },
+    );
+    const messages = Array.isArray(res?.messages) ? res.messages : [];
+    const parsed = messages
+      .map((msg) => extractMessageText(msg as ChatMessage))
+      .filter((msg): msg is { role: string; text: string } => Boolean(msg));
+    const recent = parsed.slice(-maxMessages);
+    if (recent.length === 0) {
+      return [];
+    }
+    const lines: string[] = [];
+    let total = 0;
+    for (const entry of recent) {
+      const label = entry.role === "user" ? "User" : "Assistant";
+      const text = truncateText(entry.text, REMINDER_CONTEXT_PER_MESSAGE_MAX);
+      const line = `- ${label}: ${text}`;
+      total += line.length;
+      if (total > REMINDER_CONTEXT_TOTAL_MAX) {
+        break;
+      }
+      lines.push(line);
+    }
+    return lines;
+  } catch {
+    return [];
+  }
+}
